requirements-prod.txt
# Production requirements for Zypherus
# Use this for Docker and WSGI deployments

# Core ML/NLP
numpy>=1.24.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
torch>=2.0.0
transformers>=4.25.0
pandas>=1.5.0

# Network
requests>=2.28.0
aiohttp>=3.8.0

# Web framework & security
flask>=2.0.0
flask-cors>=4.0.0
gunicorn>=20.0.0
python-dotenv>=0.19.0
Werkzeug>=2.2.0
PyJWT>=2.6.0

# Data validation
pydantic>=1.10.0

# Processing
psutil>=5.9.0

# Logging
python-json-logger>=2.0.0

# Monitoring & error tracking
sentry-sdk>=1.5.0

# Development (optional - comment out for production)
# pytest>=7.0.0
# pytest-asyncio>=0.20.0
# black>=22.0.0
# flake8>=4.0.0
